complete transfer flow

See merge request dev-4/ofogh/mobilebank-redesign-2024/mobile_bank_mellat_android!12

+++ transfer/src/main/java/com/pmb/transfer/domain/entity/TransferMethodEntity.kt
@@ -5,7 +5,7 @@ import com.pmb.core.platform.DomainModel
 data class TransferMethodEntity(
     val title: String,
     val detail: String,
-    val fee: Long,
+    val fee: Double,
     val active: Boolean,
     val default: Boolean = false,
     val paymentType: PaymentType
